LayoutTests/js/function-declarations-in-switch-statement-expected.txt

WARN: shouldBe() expects string arguments
PASS 20 is 20
WARN: shouldBe() expects string arguments
PASS 20 is 20
WARN: shouldBe() expects string arguments
PASS -1 is -1
PASS successfullyParsed is true

TEST COMPLETE
